requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b3
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.10.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
psycopg2-binary==2.9.5
pydantic==1.10.4
Pygments==2.14.0
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
yarl==1.8.2
